fix(tileEngine): fix Safari not drawing when canvas is larger than image
* Use temporary image for tileEngine renders

On Safari 12.1.2 copying the offscreenCanvas into the real canvas doesn't seem to work. The offscreenCanvas is created correctly (as can be seen through the canvas debug tool) but when the render function is drawing the image that is in it, it has no effect. Creating a new image with the content of the offscreenCanvas and using that one instead seems to fix the problem.

* Use different approach on canvas size calculation

* Use sWidth and sHeight